+++ lua/jokers/heavenlykitsune.lua
@@ -21,37 +21,21 @@ SMODS.Joker {
     calculate = function(self, card, context)
         -- Ace Conversion
         if context.before and context.cardarea == G.jokers then
-            local trigger = false
             for k, scored_card in ipairs(context.scoring_hand) do
                 if scored_card:get_id() ~= 14 and SMODS.pseudorandom_probability(card, 'tss_kitsune', 1, card.ability.extra.ace_odds) then
-                    trigger = true
-                    SMODS.change_base(scored_card, nil, 'Ace')      --Turns card into Ace
+                    SMODS.change_base(scored_card, nil, 'Ace')
                     scored_card:juice_up(0.3,0.4)
+                    SMODS.calculate_effect( { message = localize('k_poof'), colour = G.C.SECONDARY_SET.Planet }, card )
                 end
-            end
-            if trigger then
-                trigger = false
-                return {
-                    message = localize('k_poof'),                   --Shows message under Joker
-                    colour = G.C.SECONDARY_SET.Planet
-                }
             end
         end
         -- Seal Application
         if context.after and context.cardarea == G.jokers then
-            local trigger = false
             for k, scored_card in ipairs(context.scoring_hand) do
                 if scored_card:get_id() == 14 and SMODS.pseudorandom_probability(card, 'tss_kitsune', 1, card.ability.extra.fave_odds) then
-                    trigger = true
-                    scored_card:set_seal("tss_favour")          --Applies Kit's Favor        
+                    scored_card:set_seal("tss_favour")
+                    SMODS.calculate_effect( { message = localize('k_kiss'), colour = G.C.SECONDARY_SET.Planet }, card )        
                 end
-            end
-            if trigger then
-                trigger = false
-                return {
-                    message = localize('k_kiss'),                   --Shows message under Joker
-                    colour = G.C.SECONDARY_SET.Planet
-                }
             end
         end
     end